requirements.txt
streamlit==1.39.0
django-environ==0.11.2
requests==2.32.3
moviepy==1.0.3
deepl==1.19.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
streamlit_js==1.0.8
scikit-learn==1.5.2
jieba==0.42.1
